Fixes import of Admin SDK API clients
If a client is being discovered that has a version that doesn't start
with the letter v and a number, a sub-object is created.  This allows
multiple APIs with the same name but different canonical names to be
imported (such as the Admin SDK APIs).

Example:  When loading the Admin Directory API, it will be placed at
api.admin.directory instead of at api.admin.  This allows the Admin
Reporting API to be loaded at api.admin.reports.