Rework the PKCS12-to-RSA conversion example
Added the -nodes switch to disable passphrase for the pem certificate and also merged the commands together to make the conversion a single-step process. Also, since we are converting the file to key.pem, I updated the JWT constructor's certificate path to be key.pem instead of private_key.pem .

See discussion in #154.