Add --save to installation instructions
Hey. Just reading through the README I noticed that you specify that you can install the module as a dependency with the command `npm install googleapis`. This is not strictly true. It will not be added as a dependecy unless you also add --save. 

Other than that, thanks so much for making this available and have a great weekend!