Example using JWT auth client with .json files
I'm guessing that when this documentation was first written, either the Google Developer Console didn't yet provide .json files or the JWT auth client didn't yet accept keys as strings. Now that they do, it's much easier to configure the client using the .json file than it is to convert a .p12 file to .pem file. Also, the instructions for using openssl to convert a .p12 file didn't work on my system and I had to search for the correct way to do that. So, I removed the convert .p12 to .pem instructions and example and replaced them with an example using a .json file. I also added a link to the example/jwt.js in case someone wants to still use .pem files, use the impersonate functionality or any of the other functionality.