Fix wrong documentation for generateAuthUrl() (#817)
Hi everyone!
The `state` parameter in the auth URL should be a string, as documented here https://developers.google.com/identity/protocols/OAuth2WebServer .
In this README was stated that is possible to pass the `state` as an object literal, but this doesn't work. Changing the parameter to a string works like expected.