chore: update dependencies to support workload identity federation
Bump up versions for:
- `google-auth-library` to `^7.0.2`
- `googleapis-common` to `^5.0.1`

The changes above are not sufficient as the `BaseExternalAccountClient` type is still missing in the generated APIs. As a result `BaseExternalAccountClient` was added to the API templates `api-endpoint.njk` and all the APIs were regenerated.
While we're at it, `googleapis-common` dependency in the generated APIs was updated to `^5.0.1`.